Move all files out of the Roles folder
