+++ ui/src/SubHeader/SortingAndFiltering/RolesFilter/MyRolesForm/MyRolesForm.test.tsx
@@ -19,12 +19,12 @@ import React from 'react';
 import {renderWithRecoil} from 'Utils/TestUtils';
 import TestData from 'Utils/TestData';
 import {findByTestId, findByText, fireEvent, screen, waitFor} from '@testing-library/react';
-import RoleClient from '../Services/Api/RoleClient';
+import RoleClient from 'Services/Api/RoleClient';
 import MyRolesForm from './MyRolesForm';
 import ColorClient from 'Services/Api/ColorClient';
 import {RolesState} from 'State/RolesState';
 import {CurrentSpaceState} from 'State/CurrentSpaceState';
-import {RoleTagRequest} from '../Types/TagRequest';
+import {RoleTagRequest} from 'Types/TagRequest';
 
 jest.mock('Services/Api/RoleClient');
 jest.mock('Services/Api/ColorClient');
